requirements.txt
face-recognition==1.3.0
face-recognition-models==0.3.0
imageio==2.8.0
imageio-ffmpeg==0.4.1
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.0
Pillow==7.0.0
PyYAML==5.3
scikit-learn==0.22.1
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
